+++ lib/src/event.dart
@@ -1,12 +1,33 @@
 part of fletch;
 
-typedef void _FletchEventListener(Event event, Element self);
+/**
+ * The signature for event listeners handled by Fletch.
+ *
+ * [event] is a reference to the fired event, and [self] is
+ * a reference to the element owning the listener.
+ */
+typedef void FletchEventListener(Event event, Element self);
 
-class _EventFunctor {
+abstract class EventSubscription {
+    final Iterable<StreamSubscription> _subscriptions;
+
+    EventSubscription._new(this._subscriptions);
+
+    void cancel() {
+        for (var subscription in _subscriptions)
+            subscription.cancel();
+    }
+}
+
+class _EventSubscription extends EventSubscription {
+    _EventSubscription(subscriptions) : super._new(subscriptions);
+}
+
+abstract class EventFunctor {
     final String _event;
     final Iterable<Element> _elements;
 
-    _EventFunctor(this._event, this._elements);
+    EventFunctor._new(this._event, this._elements);
 
     /**
      * Dispatches the event on all selected elements.
@@ -21,15 +42,19 @@ class _EventFunctor {
      *
      * For multiple listeners, call this method multiple times.
      */
-    void listen(_FletchEventListener listener) {
+    EventSubscription listen(FletchEventListener listener) {
+        var subs = [];
         for (var element in _elements)
-            element.on[_event].listen((evt) => listener(evt, element));
+            subs.add(element.on[_event].listen((evt) => listener(evt, element)));
+        return new _EventSubscription(subs);
     }
-
-    // TODO: Unsubscribe from events
 }
 
-class _EventCollection {
+class _EventFunctor extends EventFunctor {
+    _EventFunctor(event, elements) : super._new(event, elements);
+}
+
+abstract class EventCollection {
     Fletch _parent;
 
     /**
@@ -37,76 +62,78 @@ class _EventCollection {
      *
      * The returned object can be used to dispatch and listen for the event.
      */
-    _EventFunctor operator[](String event) => new _EventFunctor(event, _parent);
+    EventFunctor operator[](String event) => new _EventFunctor(event, _parent);
 }
+
+class _EventCollection extends EventCollection {}
 
 abstract class _EventMixin {
     List<Element> _elements;
 
     /** Shorthand for `event["click"]` */
-    _EventFunctor get click       => new _EventFunctor("click",     _elements);
+    EventFunctor get click       => new _EventFunctor("click",     _elements);
 
     /** Shorthand for `event["resize"]` */
-    _EventFunctor get resize      => new _EventFunctor("resize",    _elements);
+    EventFunctor get resize      => new _EventFunctor("resize",    _elements);
 
     /** Shorthand for `event["scroll"]` */
-    _EventFunctor get scroll      => new _EventFunctor("scroll",    _elements);
+    EventFunctor get scroll      => new _EventFunctor("scroll",    _elements);
 
     /** Shorthand for `event["load"]` */
-    _EventFunctor get load        => new _EventFunctor("load",      _elements);
+    EventFunctor get load        => new _EventFunctor("load",      _elements);
 
     /** Shorthand for `event["unload"]` */
-    _EventFunctor get unload      => new _EventFunctor("unload",    _elements);
+    EventFunctor get unload      => new _EventFunctor("unload",    _elements);
 
     /** Shorthand for `event["blur"]` */
-    _EventFunctor get blur        => new _EventFunctor("blur",      _elements);
+    EventFunctor get blur        => new _EventFunctor("blur",      _elements);
 
     /** Shorthand for `event["change"]` */
-    _EventFunctor get change      => new _EventFunctor("change",    _elements);
+    EventFunctor get change      => new _EventFunctor("change",    _elements);
 
     /** Shorthand for `event["focus"]` */
-    _EventFunctor get focus       => new _EventFunctor("focus",     _elements);
+    EventFunctor get focus       => new _EventFunctor("focus",     _elements);
 
     /** Shorthand for `event["focusin"]` */
-    _EventFunctor get focusIn     => new _EventFunctor("focusin",   _elements);
+    EventFunctor get focusIn     => new _EventFunctor("focusin",   _elements);
 
     /** Shorthand for `event["focusOut"]` */
-    _EventFunctor get focusOut    => new _EventFunctor("focusout",  _elements);
+    EventFunctor get focusOut    => new _EventFunctor("focusout",  _elements);
 
     /** Shorthand for `event["select"]` */
-    _EventFunctor get select      => new _EventFunctor("select",    _elements);
+    EventFunctor get select      => new _EventFunctor("select",    _elements);
 
     /** Shorthand for `event["submit"]` */
-    _EventFunctor get submit      => new _EventFunctor("submit",    _elements);
+    EventFunctor get submit      => new _EventFunctor("submit",    _elements);
 
     /** Shorthand for `event["keydown"]` */
-    _EventFunctor get keyDown     => new _EventFunctor("keydown",   _elements);
+    EventFunctor get keyDown     => new _EventFunctor("keydown",   _elements);
 
     /** Shorthand for `event["keypress"]` */
-    _EventFunctor get keyPress    => new _EventFunctor("keypress",  _elements);
+    EventFunctor get keyPress    => new _EventFunctor("keypress",  _elements);
 
     /** Shorthand for `event["keyup"]` */
-    _EventFunctor get keyUp       => new _EventFunctor("keyUp",     _elements);
+    EventFunctor get keyUp       => new _EventFunctor("keyUp",     _elements);
 
     /** Shorthand for `event["dblclick"]` */
-    _EventFunctor get doubleClick => new _EventFunctor("dblclick",  _elements);
+    EventFunctor get doubleClick => new _EventFunctor("dblclick",  _elements);
 
     /** Shorthand for `event["hover"]` */
-    _EventFunctor get hover       => new _EventFunctor("hover",     _elements);
+    EventFunctor get hover       => new _EventFunctor("hover",     _elements);
 
     /** Shorthand for `event["mousedown"]` */
-    _EventFunctor get mouseDown   => new _EventFunctor("mousedown", _elements);
+    EventFunctor get mouseDown   => new _EventFunctor("mousedown", _elements);
 
     /** Shorthand for `event["mousemove"]` */
-    _EventFunctor get mouseMove   => new _EventFunctor("mousemove", _elements);
+    EventFunctor get mouseMove   => new _EventFunctor("mousemove", _elements);
 
     /** Shorthand for `event["mouseout"]` */
-    _EventFunctor get mouseOut    => new _EventFunctor("mouseout",  _elements);
+    EventFunctor get mouseOut    => new _EventFunctor("mouseout",  _elements);
 
     /** Shorthand for `event["mouseover"]` */
-    _EventFunctor get mouseOver   => new _EventFunctor("mouseover", _elements);
+    EventFunctor get mouseOver   => new _EventFunctor("mouseover", _elements);
 
     /** Shorthand for `event["mouseup"]` */
-    _EventFunctor get mouseUp     => new _EventFunctor("mouseup",   _elements);
+    EventFunctor get mouseUp     => new _EventFunctor("mouseup",   _elements);
 }