+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ## 0.2.0
 _Not Yet Released_
-* Added `.html` for getting/setting HTML content
+- Added `.html` for getting/setting HTML content
+- Event listeners can now be removed
 
 ## 0.1.0
 _Released 7/15/2014_
-* First release!
+- First release!